src/orchestration.py
```python
# Importing necessary libraries
from airflow import DAG
from airflow.operators.python_operator import PythonOperator
from datetime import datetime
from kfp import dsl
from kfp import compiler
from nifi import RemoteProcessGroup, NiFi

class Orchestration:
    def __init__(self, model):
        self.model = model

    def airflow_orchestration(self, tasks, dependencies):
        """
        Function to orchestrate tasks using Apache Airflow
        """
        # Define the DAG
        dag = DAG('model_pipeline', description='A pipeline to train and deploy AI model',
                  schedule_interval='0 12 * * *',
                  start_date=datetime(2021, 1, 1), catchup=False)

        # Define the tasks
        task_operators = {task_name: PythonOperator(task_id=task_name, python_callable=task_func, dag=dag) for task_name, task_func in tasks.items()}

        # Define the dependencies
        for task_name, task_dependencies in dependencies.items():
            for task_dependency in task_dependencies:
                task_operators[task_name].set_upstream(task_operators[task_dependency])

    def kubeflow_orchestration(self, pipeline_func):
        """
        Function to orchestrate tasks using Kubeflow Pipelines
        """
        # Compile the pipeline
        pipeline_func = dsl.pipeline(
            name='model_pipeline',
            description='A pipeline to train and deploy AI model'
        )(pipeline_func)
        compiler.Compiler().compile(pipeline_func, 'model_pipeline.zip')

    def nifi_orchestration(self, host, port, process_group_id):
        """
        Function to orchestrate tasks using Apache NiFi
        """
        # Connect to NiFi
        nifi = NiFi(host=host, port=port)

        # Get the process group
        process_group = nifi.pg.get(process_group_id)

        # Start the process group
        process_group.schedule_components(state='RUNNING')
```
