requirements.txt
# Explainability and Interpretability
shap==0.39.0
lime==0.2.0.1
captum==0.3.1
interpret==0.2.4

# Fairness and Bias Mitigation
fairlearn==0.6.2
aif360==0.4.0
themis-ml==0.0.4
audit-ai==0.1.3

# Monitoring and Alerting
prometheus-client==0.9.0
grafana-api==1.0.2
datadog==0.41.0
newrelic==6.2.0.156

# Debugging and Profiling
tensorboard==2.4.1
torch==1.8.1
tensorflow==2.4.1
mlflow==1.15.0

# Testing and Validation
tensorflow-data-validation==0.30.0
great-expectations==0.13.19
tfx==0.30.0
mlflow==1.15.0

# Governance and Compliance
databricks-cli==0.14.3
datarobot==2.24.0

# Collaboration and Version Control
GitPython==3.1.14

# Experimentation and Evaluation
google-cloud-aiplatform==0.7.1
dvc==2.0.18
h2o==3.32.1.3
mlflow==1.15.0

# Documentation Platforms
Sphinx==3.5.4
notebook==6.3.0

# Ethics and Compliance
# No specific Python packages

# Performance Optimization
# No specific Python packages

# Security and Access Control
boto3==1.17.49
azure-identity==1.6.0
google-auth==1.30.0
hvac==0.10.14

# Deployment and Serving
tensorflow-serving-api==2.4.1
onnxruntime==1.7.0
tritonclient==2.9.0

# Orchestration and Workflow
apache-airflow==2.0.1
kfp==1.4.0
nifi==1.13.2

# State Management and Versioning
# No specific Python packages

# Testing and Validation Strategies
# No specific Python packages

# Governance and Compliance Frameworks
# No specific Python packages

# Collaboration and Teamwork
# No specific Python packages

# Experimentation and Evaluation Best Practices
# No specific Python packages

# Documentation and Knowledge Sharing
# No specific Python packages

# Ethics and Responsible AI
# No specific Python packages

# Performance Optimization Strategies
# No specific Python packages

# Security and Access Control
# No specific Python packages
